EVG-17158: Switch from using iterator interface to returning channels.
This updates the code in retrieval to return channels of LogLines
instead of using the iterator interface.
